Update EX1.cpp
The code calculates the value of a mathematical series known as the Leibniz formula for π/4. This formula states that π/4 can be approximated by the alternating series:

1 - 1/3 + 1/5 - 1/7 + 1/9 - ...

The code prompts the user to enter a value for n, which determines the number of terms in the series. It then iterates over the odd numbers from 1 to n, alternately adding or subtracting the reciprocal of each odd number to the variable p. Finally, it outputs the value of p, which is an approximation of π/4.